payment: janitor: make IncompletePaymentAttemptTask process payment events
This is part of a series of patches consolidating IncompletePaymentAttemptTask
and IncompletePaymentTransactionTask to fix https://github.com/killbill/killbill/issues/1061.
