+++ payment/src/main/java/org/killbill/billing/payment/core/janitor/CompletionTaskBase.java
@@ -1,6 +1,6 @@
 /*
- * Copyright 2014-2017 Groupon, Inc
- * Copyright 2014-2017 The Billing Project, LLC
+ * Copyright 2014-2019 Groupon, Inc
+ * Copyright 2014-2019 The Billing Project, LLC
  *
  * The Billing Project licenses this file to you under the Apache License, version 2.0
  * (the "License"); you may not use this file except in compliance with the
@@ -19,13 +19,19 @@ package org.killbill.billing.payment.core.janitor;
 
 import java.io.IOException;
 import java.util.Iterator;
+import java.util.List;
+import java.util.UUID;
 
+import javax.annotation.Nullable;
+
+import org.joda.time.DateTime;
 import org.killbill.billing.account.api.AccountApiException;
 import org.killbill.billing.account.api.AccountInternalApi;
 import org.killbill.billing.account.api.ImmutableAccountData;
 import org.killbill.billing.callcontext.DefaultCallContext;
 import org.killbill.billing.callcontext.InternalTenantContext;
 import org.killbill.billing.events.PaymentInternalEvent;
+import org.killbill.billing.payment.api.TransactionStatus;
 import org.killbill.billing.payment.core.sm.PaymentControlStateMachineHelper;
 import org.killbill.billing.payment.core.sm.PaymentStateMachineHelper;
 import org.killbill.billing.payment.dao.PaymentDao;
@@ -41,13 +47,22 @@ import org.killbill.clock.Clock;
 import org.killbill.commons.locker.GlobalLock;
 import org.killbill.commons.locker.GlobalLocker;
 import org.killbill.commons.locker.LockFailedException;
+import org.killbill.notificationq.api.NotificationEvent;
 import org.killbill.notificationq.api.NotificationQueue;
+import org.skife.config.TimeSpan;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
 
 abstract class CompletionTaskBase<T> implements Runnable {
 
     private static final Logger log = LoggerFactory.getLogger(CompletionTaskBase.class);
+
+    protected static final ImmutableList<TransactionStatus> TRANSACTION_STATUSES_TO_CONSIDER = ImmutableList.<TransactionStatus>builder().add(TransactionStatus.PENDING)
+                                                                                                                                         .add(TransactionStatus.UNKNOWN)
+                                                                                                                                         .build();
 
     protected final PaymentConfig paymentConfig;
     protected final Clock clock;
@@ -122,8 +137,6 @@ abstract class CompletionTaskBase<T> implements Runnable {
 
     public abstract void doIteration(final T item);
 
-    public abstract void processPaymentEvent(final PaymentInternalEvent event, final NotificationQueue janitorQueue) throws IOException;
-
     public void attachJanitorQueue(final NotificationQueue janitorQueue) {
         this.janitorQueue = janitorQueue;
     }
@@ -160,5 +173,46 @@ abstract class CompletionTaskBase<T> implements Runnable {
     protected CallContext createCallContext(final String taskName, final InternalTenantContext internalTenantContext) {
         final TenantContext tenantContext = internalCallContextFactory.createTenantContext(internalTenantContext);
         return new DefaultCallContext(null, tenantContext.getTenantId(), taskName, CallOrigin.INTERNAL, UserType.SYSTEM, UUIDs.randomUUID(), clock);
+    }
+
+    protected void insertNewNotificationForUnresolvedTransactionIfNeeded(final UUID paymentTransactionId, final TransactionStatus transactionStatus, @Nullable final Integer attemptNumber, @Nullable final UUID userToken, final Long accountRecordId, final Long tenantRecordId) {
+        // When we come from a GET path, we don't want to insert a new notification
+        if (attemptNumber == null) {
+            return;
+        }
+
+        final InternalTenantContext tenantContext = internalCallContextFactory.createInternalTenantContext(tenantRecordId, accountRecordId);
+
+        // Increment value before we insert
+        final Integer newAttemptNumber = attemptNumber + 1;
+        final NotificationEvent key = new JanitorNotificationKey(paymentTransactionId, IncompletePaymentTransactionTask.class.toString(), newAttemptNumber);
+        final DateTime notificationTime = getNextNotificationTime(transactionStatus, newAttemptNumber, tenantContext);
+        // Will be null in the GET path or when we run out opf attempts..
+        if (notificationTime != null) {
+            try {
+                janitorQueue.recordFutureNotification(notificationTime, key, userToken, accountRecordId, tenantRecordId);
+            } catch (final IOException e) {
+                log.warn("Failed to insert future notification for paymentTransactionId = {}: {}", paymentTransactionId, e.getMessage());
+            }
+        }
+    }
+
+    @VisibleForTesting
+    DateTime getNextNotificationTime(final TransactionStatus transactionStatus, final Integer attemptNumber, final InternalTenantContext internalTenantContext) {
+        final List<TimeSpan> retries;
+        if (TransactionStatus.UNKNOWN.equals(transactionStatus)) {
+            retries = paymentConfig.getUnknownTransactionsRetries(internalTenantContext);
+        } else if (TransactionStatus.PENDING.equals(transactionStatus)) {
+            retries = paymentConfig.getPendingTransactionsRetries(internalTenantContext);
+        } else {
+            retries = ImmutableList.of();
+            log.warn("Unexpected transactionStatus='{}' from janitor, ignore...", transactionStatus);
+        }
+
+        if (attemptNumber > retries.size()) {
+            return null;
+        }
+        final TimeSpan nextDelay = retries.get(attemptNumber - 1);
+        return clock.getUTCNow().plusMillis((int) nextDelay.getMillis());
     }
 }
